BIG CHANGE: each site to a specific file, and main cleaned.
