--- old_links.txt
@@ -1,10 +1,0 @@
-
-https://www.jdsports.co.il/products/nicn5668101
-https://www.jdsports.co.il/products/nidr0453001
-https://www.jdsports.co.il/products/nidz4514300
-https://www.jdsports.co.il/products/nifd0654100
-https://www.jdsports.co.il/products/jodq0560160
-https://www.jdsports.co.il/products/nidj9779001
-https://www.jdsports.co.il/products/nidz4504300
-https://www.jdsports.co.il/products/nidq3984002
-
